refactor: update UserButton component
