--- src/Layout/Header/modules/HeaderStickyBar/components/UserButton/UserButton.module.scss
@@ -1,11 +1,0 @@
-@use '@/styles/colors.scss';
-
-.button {
-  width: 50px;
-  height: 50px;
-
-  font-weight: 500;
-
-  background-color: colors.$isabelline;
-  border-radius: 50%;
-}
