+++ src/Layout/Header/modules/UserActions/components/UserButton/types.ts
@@ -1,0 +1,4 @@
+export type UserButtonProps = {
+  firstName: string;
+  lastName: string;
+};
